feat(clipboard): Implement categories, drag-and-drop, and settings module
This commit introduces category management for clipboard history, a drag-and-drop interface for organizing items, and a dedicated settings module. It also refactors database identifiers and window management for the desktop application.

### Key Changes:

- **Categories Feature:**
  - Introduced `CategoryEntity` and `CategoryCopiedDataCrossRef` in the database to support organization of clipboard items.
  - Added a new `shared/features/setting` module with `SettingScreen` and `CategorySetting` to manage (add, rename, delete) categories.
  - Updated `ClipboardViewModel` to handle category selection, filtering, and data observation.

- **Drag and Drop:**
  - Implemented a generic `DragAndDropContainer` and `DragAndDropState` in `shared/core/ui`.
  - Enabled dragging clipboard items into specific categories within the `ClipboardScreen`.

- **Database & Domain Refactor:**
  - Migrated primary keys for `CopiedData` from `Int` to `Long` across the database, domain entities, and repositories.
  - Added `CategoryDao` and updated `LocalClipboardRepository` to support category CRUD operations and relationships.
  - Introduced new use cases: `AddCategoryUseCase`, `AddCopiedDataToCategoryUseCase`, `ObserveCopiedDataByCategoryUseCase`, and others.

- **UI & Window Management:**
  - Refactored desktop window logic into `QuickClipboardWindow` (overlay) and `GeneralWindow` (settings).
  - Added double-click functionality to clipboard items for immediate pasting.
  - Created reusable UI components: `TopBar`, `RoundedIconButton`, and new icons (`Edit`, `Plus`, `Trash`).
  - Updated `DesktopClipboardManager` to use `Long` keys for the LRU cache.

+++ desktopApp/src/main/kotlin/DesktopClipboardManager.kt
@@ -40,17 +40,17 @@ class DesktopClipboardManager() : ClipboardManager {
 
     // Synchronized LRU cache for clipboard data
     private val savedCopiedDataLock = Any()
-    private val savedCopiedData = object : LinkedHashMap<Int, CopiedData>(256, 0.75f, true) {
-        override fun removeEldestEntry(eldest: MutableMap.MutableEntry<Int, CopiedData>?): Boolean {
+    private val savedCopiedData = object : LinkedHashMap<Long, CopiedData>(256, 0.75f, true) {
+        override fun removeEldestEntry(eldest: MutableMap.MutableEntry<Long, CopiedData>?): Boolean {
             return size > 256
         }
     }
 
-    private fun cachePut(hash: Int, value: CopiedData) {
+    private fun cachePut(hash: Long, value: CopiedData) {
         synchronized(savedCopiedDataLock) { savedCopiedData[hash] = value }
     }
 
-    private fun cacheGet(hash: Int): CopiedData? =
+    private fun cacheGet(hash: Long): CopiedData? =
         synchronized(savedCopiedDataLock) { savedCopiedData[hash] }
 
     private val coroutineScope = CoroutineScope(Dispatchers.IO)
@@ -219,7 +219,7 @@ class DesktopClipboardManager() : ClipboardManager {
                 }
                 if (!list.isNullOrEmpty()) {
                     val paths = list.map { it.absolutePath }
-                    val hash = computeFilesHash(paths, list)
+                    val hash = computeFilesHash(paths, list).toLong()
                     val existing = cacheGet(hash)
                     if (existing == null) {
                         val fileData = CopiedData.File(
@@ -249,7 +249,7 @@ class DesktopClipboardManager() : ClipboardManager {
                     null
                 }
                 if (image != null) {
-                    val hash = imageHashDownscaled(image, 128)
+                    val hash = imageHashDownscaled(image, 128).toLong()
                     if (cacheGet(hash) == null) {
                         val cacheDir = File(System.getProperty("user.home"), ".crosssync/images")
                         val fileName = "${hash}.png"
@@ -341,7 +341,7 @@ class DesktopClipboardManager() : ClipboardManager {
             }
 
             if (!formattedRaw.isNullOrEmpty() && mimeType != null && !plainNormalized.isNullOrEmpty()) {
-                val hash = computeStringHash(plainNormalized, mimeType)
+                val hash = computeStringHash(plainNormalized, mimeType).toLong()
                 val existing = cacheGet(hash)
                 if (existing == null) {
                     val data = CopiedData.FormattedText(
@@ -385,7 +385,7 @@ class DesktopClipboardManager() : ClipboardManager {
                 null
             }
             if (text.isNullOrEmpty()) return null
-            val hash = computeStringHash(text, "text/plain")
+            val hash = computeStringHash(text, "text/plain").toLong()
             val existing = cacheGet(hash)
             if (existing == null) {
                 val found = synchronized(savedCopiedDataLock) {
